requirements.txt
numpy~=2.0.2
pandas~=2.2.3
PyYAML==6.0.2

fpdf~=1.7.2
selenium~=4.28.1